lift types into `types.ts`
